+++ CHANGELOG.md
@@ -4,11 +4,23 @@ All notable changes to the "rush-vscode" extension will be documented in this fi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [0.0.2]
+## 0.1.0
+
+### Features
+
+- Added "Rush: Start package" command to run "start" script of specified package
+
+- Added inline buttons for "Start package" and "Build dependencies" to tree view
+
+### Bugfixes
+
+- Fixed command unavailability when tree view is not expanded
+
+## 0.0.2
 
 - Added extension icon
 - Project dependencies in tree view are grouped into folder
 
-## [0.0.1]
+## 0.0.1
 
 - Initial release